requirements.txt
requests
pyyaml
azure-cli
